Adds dotnet test reporter
Adds the dotnet-test-reporter action to generate test reports, but only if SonarQube analysis is disabled via the ENABLE_SONAR environment variable.
This ensures that test reports are generated in environments where SonarQube is not used for analysis, providing comprehensive test coverage insights.

Relates to CLOUD-2775

Adds debugging and test flexibility

Adds debugging commands to the test script to print the current directory and content of coverage files.

Allows failed tests when the 'wait_flag' environment variable is not set, providing flexibility in test execution based on environment configuration.

Relates to CLOUD-2775

[CLOUD-2775]: Generates and uploads coverage reports

Adds the generation of pretty HTML and Markdown summary coverage reports using the ReportGenerator GitHub Action.

It also uploads the generated coverage report as an artifact for easier access and analysis.

The changes include disabling the dotnet-test-reporter when SonarQube analysis is enabled and adds Release configuration to build and test commands.